decision_tree_builder/requirements.txt
Flask>=2.3,<3.0
PyYAML>=6.0
networkx>=3.1
Pillow>=9.0
